refactor truncateDuration into common functionality
