+++ lib/animas/animasSimulator.js
@@ -20,6 +20,7 @@ var sundial = require('sundial');
 var util = require('util');
 
 var annotate = require('../eventAnnotations');
+var common = require('../commonFunctions');
 
 var debug = require('../bows')('AnimasDriver');
 
@@ -74,21 +75,6 @@ exports.make = function(config){
     return e;
   }
 
-  function truncateDuration(basal) {
-    var fiveDays = (5 * 1440 * sundial.MIN_TO_MSEC);
-    if(basal.isAssigned('duration')) {
-      if(basal.duration > fiveDays) {
-        //flat-rate basal
-        basal.duration = fiveDays;
-        annotate.annotateEvent(basal, 'animas/basal/flat-rate');
-      }
-    } else {
-      basal.with_duration(0);
-      annotate.annotateEvent(basal, 'basal/unknown-duration');
-    }
-    return basal;
-  }
-
   function simpleSimulate(e) {
     ensureTimestamp(e);
     events.push(e);
@@ -114,7 +100,7 @@ exports.make = function(config){
           var duration = Date.parse(event.time) - Date.parse(currBasal.time);
           currBasal.with_duration(duration);
         }
-        currBasal = truncateDuration(currBasal);
+        currBasal = common.truncateDuration(currBasal, 'animas');
       }
 
       if (event.rate === 0 ) {
@@ -181,7 +167,7 @@ exports.make = function(config){
       if(currBasal != null && currBasal.rate !== 0 && currStatus.status === 'suspended') {
         //basal rate has not yet changed to zero after final suspend event
         currBasal.with_duration(Date.parse(currStatus.time) - Date.parse(currBasal.time));
-        currBasal = truncateDuration(currBasal);
+        currBasal = common.truncateDuration(currBasal, 'animas');
         currBasal = currBasal.done();
         events.push(currBasal);
       }
