+++ lib/commonFunctions.js
@@ -38,6 +38,21 @@ exports.computeMillisInCurrentDay = function(e){
   return msFromMidnight;
 };
 
+exports.truncateDuration = function(basal, source) {
+  var fiveDays = (5 * 1440 * sundial.MIN_TO_MSEC);
+  if(basal.isAssigned('duration')) {
+    if(basal.duration > fiveDays) {
+      //flat-rate basal
+      basal.duration = fiveDays;
+      annotate.annotateEvent(basal, source + '/basal/flat-rate');
+    }
+  } else {
+    basal.with_duration(0);
+    annotate.annotateEvent(basal, 'basal/unknown-duration');
+  }
+  return basal;
+};
+
 exports.bytes2hex = function(bytes, noGaps) {
   var message = '';
   for(var i in bytes) {
